# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/plans": {
      "post": {
        "operationId": "PlanController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new plan",
        "tags": [
          "Plans"
        ]
      },
      "get": {
        "operationId": "PlanController_findAll",
        "parameters": [
          {
            "name": "filter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plans have been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all plans",
        "tags": [
          "Plans"
        ]
      }
    },
    "/plans/{id}": {
      "delete": {
        "operationId": "PlanController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Plan has been successfully deleted."
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a plan",
        "tags": [
          "Plans"
        ]
      }
    },
    "/students": {
      "post": {
        "operationId": "StudentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Student data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The student has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a student",
        "tags": [
          "Students"
        ]
      },
      "get": {
        "operationId": "StudentController_findAllForCurrentTeacher",
        "parameters": [
          {
            "name": "filter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacher_id",
            "required": false,
            "in": "query",
            "description": "Teacher ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Students have been successfully retrieved for current teacher",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all students for current teacher, or if current teacher is admin, get students by teacher ID in query",
        "tags": [
          "Students"
        ]
      }
    },
    "/students/{id}": {
      "get": {
        "operationId": "StudentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student has been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentExtendedDto"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a student",
        "tags": [
          "Students"
        ]
      },
      "patch": {
        "operationId": "StudentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The student has been successfully updated."
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a student",
        "tags": [
          "Students"
        ]
      },
      "delete": {
        "operationId": "StudentController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The student has been successfully deleted."
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a student",
        "tags": [
          "Students"
        ]
      }
    },
...
    "/lessons/student/{student_id}": {
      "get": {
        "operationId": "LessonController_findLessonsForPeriodAndStudent",
        "parameters": [
          {
            "name": "student_id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "required": true,
            "in": "query",
            "description": "Start date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": true,
            "in": "query",
            "description": "End date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lessons have been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLessonsOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all lessons for a period for a student",
        "tags": [
          "Lessons"
        ]
      }
    },
...
    "/reports/schedule/download": {
      "get": {
        "operationId": "ReportsController_downloadSchedule",
        "parameters": [
          {
            "name": "start_date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/students/download": {
      "get": {
        "operationId": "ReportsController_downloadStudents",
        "parameters": [
          {
            "name": "filter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/salary": {
      "get": {
        "operationId": "ReportsController_getDataForSalary",
        "parameters": [
          {
            "name": "start_date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacher_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "School API",
    "description": "The School API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BadRequestErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "The HTTP status code",
            "example": 400
          },
          "path": {
            "type": "string",
            "description": "The path of the request",
            "example": "/api/v1/users"
          },
          "message": {
            "type": "object",
            "description": "The message of the error",
            "example": [
              "Error message 1",
              "Error message 2"
            ]
          }
        },
        "required": [
          "statusCode",
          "path",
          "message"
        ]
      },
...
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "plan_type": {
            "type": "string",
            "enum": [
              "INDIVIDUAL",
              "PAIR"
            ]
          },
          "plan_currency": {
            "type": "string",
            "enum": [
              "USD",
              "EUR",
              "PLN",
              "BYN"
            ]
          },
          "plan_price": {
            "type": "number",
            "example": 1000000
          },
          "duration": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "plan_type",
          "plan_currency",
          "plan_price",
          "duration"
        ]
      },
      "PlanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The id of the plan",
            "example": 1
          },
          "plan_name": {
            "type": "string",
            "description": "The name of the plan",
            "example": "10 minutes"
          },
          "plan_price": {
            "type": "number",
            "description": "The price of the plan",
            "example": 1000000
          },
          "plan_currency": {
            "type": "string",
            "description": "The currency of the plan",
            "example": "USD"
          },
          "duration": {
            "type": "number",
            "description": "The duration of the plan",
            "example": 10
          },
          "plan_type": {
            "type": "string",
            "description": "The type of the plan",
            "example": "INDIVIDUAL"
          },
          "deleted_at": {
            "type": "object",
            "description": "The deleted at of the plan",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "The created at of the plan",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "plan_name",
          "plan_price",
          "plan_currency",
          "duration",
          "plan_type",
          "deleted_at",
          "created_at"
        ]
      },
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the student",
            "example": "John Doe"
          },
          "class": {
            "type": "number",
            "description": "The class of the student",
            "example": 1
          },
          "birth_date": {
            "type": "object",
            "description": "The birth date of the student",
            "example": "2000-01-01T00:00:00.000Z"
          },
          "teacher_id": {
            "type": "number",
            "description": "The teacher id of the student",
            "example": 1
          },
          "timezone": {
            "type": "string",
            "description": "The timezone of the student",
            "example": "BY"
          }
        },
        "required": [
          "name",
          "class",
          "birth_date",
          "teacher_id",
          "timezone"
        ]
      },
      "StudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The id of the student",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the student",
            "example": "John Doe"
          },
          "class": {
            "type": "number",
            "description": "The class of the student",
            "example": 1
          },
          "birth_date": {
            "type": "object",
            "description": "The birth date of the student",
            "example": "2000-01-01"
          },
          "age": {
            "type": "object",
            "description": "The age of the student in full years",
            "example": 20,
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "The created at of the student",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "deleted_at": {
            "type": "object",
            "description": "The deleted at of the student",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "teacher_id": {
            "type": "object",
            "description": "The teacher id of the student",
            "example": 1
          },
          "timezone": {
            "type": "object",
            "description": "The timezone of the student",
            "example": "BY"
          }
        },
        "required": [
          "id",
          "name",
          "class",
          "birth_date",
          "age",
          "created_at",
          "deleted_at",
          "teacher_id",
          "timezone"
        ]
      },
      "StudentExtendedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The id of the student",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the student",
            "example": "John Doe"
          },
          "class": {
            "type": "number",
            "description": "The class of the student",
            "example": 1
          },
          "birth_date": {
            "type": "object",
            "description": "The birth date of the student",
            "example": "2000-01-01"
          },
          "age": {
            "type": "object",
            "description": "The age of the student in full years",
            "example": 20,
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "The created at of the student",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "deleted_at": {
            "type": "object",
            "description": "The deleted at of the student",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "teacher_id": {
            "type": "object",
            "description": "The teacher id of the student",
            "example": 1
          },
          "timezone": {
            "type": "object",
            "description": "The timezone of the student",
            "example": "BY"
          },
          "actualPlans": {
            "description": "The actual plans of the student",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanDto"
            }
          },
          "balance": {
            "type": "number",
            "description": "The balance of the student",
            "example": 0
          },
          "bookUntilCancellation": {
            "type": "boolean",
            "description": "The book until cancellation of the student",
            "example": false
          },
          "notifyAboutBirthday": {
            "type": "boolean",
            "description": "The notify about birthday of the student",
            "example": true
          },
          "notifyAboutLessons": {
            "type": "boolean",
            "description": "The notify about lessons of the student",
            "example": true
          }
        },
        "required": [
          "id",
          "name",
          "class",
          "birth_date",
          "age",
          "created_at",
          "deleted_at",
          "teacher_id",
          "timezone",
          "actualPlans",
          "balance",
          "bookUntilCancellation",
          "notifyAboutBirthday",
          "notifyAboutLessons"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "birth_date": {
            "type": "object",
            "description": "The birth date of the student",
            "example": "2000-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "birth_date"
        ]
      },
...
      "StudentLessonsOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Student name",
            "example": "John Doe"
          },
          "class": {
            "type": "number",
            "description": "Student class",
            "example": 1
          },
          "canceled_lessons": {
            "type": "number",
            "description": "Canceled lessons",
            "example": 1
          },
          "missed_lessons": {
            "type": "number",
            "description": "Missed lessons",
            "example": 1
          }
        },
        "required": [
          "id",
          "name",
          "class",
          "canceled_lessons",
          "missed_lessons"
        ]
      },
...
    }
  }
}